Updates style of back button for Partner form.
- Changes 'Back' link to be a 'Cancel' button.
- Removes 'Back' and 'Show' links from the bottom of the Edit page.
- Removes the scaffolded scss for visited links.

For stylistic purposes changes from having a "Back" link at the bottom of the
page to instead be formatted as red "Cancel" button. This change helps make this
option more visible.

In order for the link "button" to link like the submit button we needed to
remove the style for visited links. If a user clicked on the back/cancel button
the text would be styled differently than the Submit button. Since the `visited`
style was in a scaffold file, we assumed it was not intentional and okay to
remove.