Respond with a forbidden response if DIAPER_KEY env var is not set
DIAPER_KEY is an important variable granting access to key features of the app. If this environment variable is not it could be a symptom of a greater problem and we don't want to inadvertently grant access to unauthorized requests. It's also important to note that without this check, a nil X-Api-Key header and missing DIAPER_KEY variable results in a successful request, which should not.